Use the Data Requirements generated by the cql-to-elm translator when processing the measure's Library Resource for export.
The translator's Data Requirement Processor correctly maps the US Core or QI-Core profile types to their FHIR-based equivalent.

Using the ANTLR parser skips the mapping and results in the profile type name appearing in the Data Requirement, e.g., observation-bmi instead of Observation.

The R5 to R4 converter then errors out with the ANTLR results as it does not understand non-FHIR based types.

Signed-off-by: Joseph Kotanchik <joseph.kotanchik@semanticbits.com>